+++ source/extensions/lightspeed.trex.properties_pane.shared.asset_replacements.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.9]
+### Fixed
+- Adding layer validation tests
+
 ## [1.2.8]
 ### Fixed
 - Fixed layer validation